fix(patch/dwmc): fix quit command to be handled by signal function feat(patch/dwmc): disable ONLYQUITONEMPTY_PATCH
The `quit` command was not being properly handled by the `signal` function in the `dwmc` script. This commit fixes this issue by adding `quit` to the list of commands handled by the `signal` function.

The `ONLYQUITONEMPTY_PATCH` was enabled, which prevents dwm from quitting if there are any windows open. This commit disables this patch by setting `#define ONLYQUITONEMPTY_PATCH 0`. This allows dwm to quit even if there are windows open.